Rewrite the component to remove two-way bindings.
Instead, it follows Ember's "Data Down, Actions Up" paradigm. Any bound
properties will cause updates to the video.js player, but changes to the
player directly (e.g. when a user clicks play or pause) do not cause
changes to the component. Instead, they send actions which can be caught
by the application to update the bound properties which will, in turn,
cause the changes to be propagated back down to the player if necessary.

Note that the `autoresize` feature has been removed, since video.js
already has a `fluid` option that does essentially the same thing, but
purely in CSS.