Add a simplified `ivy-videojs` which wraps `ivy-videojs-player`.
This provides most of the functionality that existed prior to this
refactoring, but it accomplishes this by wrapping the lower-level
video.js player component, now called `ivy-videojs-player`.

Exposing the player as a separate component allows you to use it
yourself if you need lower-level access in your application. This might
be the case if, for instance, you're using video.js plugins.