Downgrade jQuery to 1.11.x to fix assertion.
Currently, Ember has assertions around the version of jQuery being used,
which fail with the newly-released jQuery 1.12. There is a fix in the
works, but it hasn't been released yet:

  https://github.com/emberjs/ember.js/pull/12793

For now I've just changed the jQuery dependency from "^1.11.3" to
"~1.11.3.", which will target the latest 1.11.x release, but will not
use 1.12.